pyproject.toml
[project]
name = "swarmai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "groq>=0.37.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.2",
    "langsmith>=0.7.17",
    "matplotlib>=3.10.8",
    "mcp>=1.26.0",
    "numpy>=2.4.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "streamlit>=1.55.0",
    "tavily-python>=0.7.23",
]
